stt-service/requirements.txt
openai-whisper
Flask==2.3.2